Use more bits of HDrand() to decide whether to re-acquire or release the lock.  The previous code would not complete on NetBSD 8.0 because of the poor randomness of the rand(3) implementation there.
Disable the r/w lock statistics comparison and discrepancy reporting.

Add a test that measures and reports the runtime for 100 million
non-recursive read-lock acquire/release cycles, for 100 million
recursive cycles, and the same for write-locks.

Measure and report the runtime for some of the existing r/w lock "smoke
checks."